Fix module layout
Resolves circular import issues
