+++ mephenger/models/model.py
@@ -1,0 +1,42 @@
+from abc import ABC, abstractmethod
+
+
+class Model(ABC):
+    @property
+    @abstractmethod
+    def json(self) -> dict:
+        """
+        Get the json representation of this object.
+        """
+        pass
+
+    @abstractmethod
+    def db_push(self):
+        """
+        Update the database's state of this object.
+
+        # Errors
+
+        Raises a `TimeoutExpired` exception if the operation takes longer than
+        the configured timeout.
+        """
+        pass
+
+    @abstractmethod
+    def db_fetch(self) -> 'Model':
+        """
+        Update this object's state from the database.
+
+        # Returns
+
+        The updated object.
+
+        # Errors
+
+        Raises a `TimeoutExpired` exception if the operation takes longer than
+        the configured timeout.
+
+        Raises a `NoSuchItem` exception if this object is not present in the
+        database. You may solve this by calling `db_push()`.
+        """
+        pass
